Move common Enums to their own file
