+++ traffic_control/models/road_marking.py
@@ -8,6 +8,15 @@ from django.db.models import Q
 from django.utils.translation import gettext_lazy as _
 from enumfields import Enum, EnumField, EnumIntegerField
 
+from traffic_control.enums import (
+    Condition,
+    DeviceTypeTargetModel,
+    InstallationStatus,
+    LaneNumber,
+    LaneType,
+    Lifecycle,
+    TrafficControlDeviceTypeType,
+)
 from traffic_control.mixins.models import (
     AbstractFileModel,
     SoftDeleteModel,
@@ -15,18 +24,7 @@ from traffic_control.mixins.models import (
     UpdatePlanLocationMixin,
     UserControlModel,
 )
-from traffic_control.models.common import (
-    Condition,
-    DeviceTypeTargetModel,
-    InstallationStatus,
-    LaneNumber,
-    LaneType,
-    Lifecycle,
-    OperationBase,
-    OperationType,
-    TrafficControlDeviceType,
-    TrafficControlDeviceTypeType,
-)
+from traffic_control.models.common import OperationBase, OperationType, TrafficControlDeviceType
 from traffic_control.models.plan import Plan
 from traffic_control.models.traffic_sign import TrafficSignPlan, TrafficSignReal
 from traffic_control.models.utils import SoftDeleteQuerySet
