+++ traffic_control/models/traffic_sign.py
@@ -9,6 +9,17 @@ from django.db.models import Q
 from django.utils.translation import gettext_lazy as _
 from enumfields import Enum, EnumField, EnumIntegerField
 
+from traffic_control.enums import (
+    Condition,
+    DeviceTypeTargetModel,
+    InstallationStatus,
+    LaneNumber,
+    LaneType,
+    Lifecycle,
+    Reflection,
+    Size,
+    Surface,
+)
 from traffic_control.mixins.models import (
     AbstractFileModel,
     DecimalValueFromDeviceTypeMixin,
@@ -18,20 +29,7 @@ from traffic_control.mixins.models import (
     UserControlModel,
 )
 from traffic_control.models.affect_area import CoverageArea
-from traffic_control.models.common import (
-    Condition,
-    DeviceTypeTargetModel,
-    InstallationStatus,
-    LaneNumber,
-    LaneType,
-    Lifecycle,
-    OperationBase,
-    OperationType,
-    Reflection,
-    Size,
-    Surface,
-    TrafficControlDeviceType,
-)
+from traffic_control.models.common import OperationBase, OperationType, TrafficControlDeviceType
 from traffic_control.models.mount import MountPlan, MountReal
 from traffic_control.models.plan import Plan
 from traffic_control.models.utils import SoftDeleteQuerySet
