+++ traffic_control/serializers/common.py
@@ -6,8 +6,8 @@ from enumfields.drf import EnumSupportSerializerMixin
 from rest_framework import serializers
 from rest_framework_gis.fields import GeometryField
 
+from traffic_control.enums import DeviceTypeTargetModel
 from traffic_control.models import OperationalArea, Owner, TrafficControlDeviceType
-from traffic_control.models.common import DeviceTypeTargetModel
 from traffic_control.schema import TrafficSignType
