+++ traffic_control/serializers/traffic_light.py
@@ -2,6 +2,7 @@ from enumfields.drf import EnumSupportSerializerMixin
 from rest_framework import serializers
 from rest_framework_gis.fields import GeometryField
 
+from traffic_control.enums import DeviceTypeTargetModel
 from traffic_control.models import (
     TrafficControlDeviceType,
     TrafficLightPlan,
@@ -9,7 +10,6 @@ from traffic_control.models import (
     TrafficLightReal,
     TrafficLightRealFile,
 )
-from traffic_control.models.common import DeviceTypeTargetModel
 from traffic_control.models.traffic_light import TrafficLightRealOperation
