+++ traffic_control/tests/management/commands/test_import_traffic_signs_vaisala.py
@@ -4,13 +4,13 @@ from unittest.mock import patch
 from django.core.management import call_command, CommandError
 from django.test import TestCase
 
+from traffic_control.enums import DeviceTypeTargetModel
 from traffic_control.models import (
     AdditionalSignContentReal,
     AdditionalSignReal,
     TrafficControlDeviceType,
     TrafficSignReal,
 )
-from traffic_control.models.common import DeviceTypeTargetModel
 from traffic_control.tests.management.commands.utils import mock_open
