+++ traffic_control/tests/test_admin.py
@@ -6,8 +6,8 @@ from django.contrib.gis.geos import Point
 from django.test import RequestFactory, TestCase
 
 from traffic_control.admin import BarrierRealAdmin, TrafficSignRealAdmin
+from traffic_control.enums import Lifecycle
 from traffic_control.models import BarrierReal, TrafficSignReal
-from traffic_control.models.common import Lifecycle
 from traffic_control.tests.factories import get_additional_sign_real, get_barrier_real, get_owner, get_user
