+++ traffic_control/tests/test_signpost_api.py
@@ -6,8 +6,8 @@ from django.urls import reverse
 from rest_framework import status
 from rest_framework_gis.fields import GeoJsonDict
 
+from traffic_control.enums import DeviceTypeTargetModel
 from traffic_control.models import SignpostPlan, SignpostReal
-from traffic_control.models.common import DeviceTypeTargetModel
 from traffic_control.tests.factories import (
     add_signpost_real_operation,
     get_api_client,
