+++ traffic_control/tests/test_traffic_light_api.py
@@ -5,8 +5,8 @@ from django.urls import reverse
 from rest_framework import status
 from rest_framework_gis.fields import GeoJsonDict
 
+from traffic_control.enums import DeviceTypeTargetModel
 from traffic_control.models import TrafficLightPlan, TrafficLightReal, TrafficLightSoundBeaconValue, TrafficLightType
-from traffic_control.models.common import DeviceTypeTargetModel
 from traffic_control.tests.factories import (
     add_traffic_light_real_operation,
     get_api_client,
