+++ traffic_control/tests/test_traffic_sign_api.py
@@ -6,8 +6,8 @@ from django.urls import reverse
 from rest_framework import status
 from rest_framework_gis.fields import GeoJsonDict
 
+from traffic_control.enums import DeviceTypeTargetModel
 from traffic_control.models import TrafficSignPlan, TrafficSignReal
-from traffic_control.models.common import DeviceTypeTargetModel
 from traffic_control.tests.factories import (
     add_traffic_sign_real_operation,
     get_api_client,
